# ci: Run docker with host network instead of bridge