--- a/misc/dbt-materialize/Dockerfile
+++ b/misc/dbt-materialize/Dockerfile
@@ -11,5 +11,5 @@ FROM python:3.10.14
 
 COPY . dbt-materialize/
 
-RUN pip install pytest
+RUN pip install pytest pytest-split
 RUN pip install ./dbt-materialize[dev]