… ci/nightly/pipeline.template.yml
… ci/release-qualification/pipeline.template.yml
… ci/test/cargo-test/mzcompose.py
… ci/test/pipeline.template.yml
… misc/python/materialize/mzcompose/services/materialized.py
… misc/python/materialize/parallel_workload/action.py
… misc/python/materialize/zippy/mz_actions.py
… src/materialized/ci/entrypoint.sh
… src/persist/src/azure.rs
… src/testdrive/src/action/sql.rs
… test/0dt/mzcompose.py
… test/balancerd/mzcompose.py
… test/cloudtest/test_privatelink_connection.py
… test/cloudtest/test_ssh_tunnels.py
… test/cluster-isolation/mzcompose.py
… test/cluster/blue-green-deployment/setup.td
… test/cluster/mzcompose.py
… test/cluster/storage/01-create-sources.td
… test/data-ingest/mzcompose.py
… test/feature-benchmark/mzcompose.py
… test/feature-flag-consistency/mzcompose.py
… test/kafka-auth/mzcompose.py
… test/kafka-auth/test-kafka-ssl.td
… test/kafka-rtr/resumption/mz-setup.td
… test/kafka-rtr/resumption/toxiproxy-setup.td
… test/kafka-rtr/simple/mz-setup.td
… test/kafka-rtr/simple/toxiproxy-setup.td
… test/limits/mzcompose.py
… test/mysql-cdc-resumption-old-syntax/alter-source-connection.td
… test/mysql-cdc-resumption-old-syntax/alter-table-fix.td