Only for CI, not locally. Some tests run twice as fast. Individual benchmarks are interesting to see how it affects some individual actions: https://docs.google.com/spreadsheets/d/1L-mAGsB5tcrIpqJDhGC13JRMP8TGw4nE6QYA-jbKt5s/edit?gid=2146535294#gid=2146535294
Impressive changes:
- ManySmallUpdates 18.6s -> 0.7s (27 times faster)
- SubscribeParallelTable: 58.5 s -> 2.7 s (21 times faster)
- StartupLoaded 3.7 s -> 1.7 s (2.1 times faster)
- PgCdcInitialLoad 1.2 s -> 0.7 s (1.8 times faster)
### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly considered. ([trigger-ci for additional test/nightly runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md), is a design doc ([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)), or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T` mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md) (possibly in a backwards-incompatible way), then it is tagged with a `T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests, there is a companion cloud PR to account for those changes that is tagged with the release-blocker label ([example](https://github.com/MaterializeInc/cloud/pull/5021)).
  <!-- Ask in #team-cloud on Slack if you need help preparing the cloud PR. -->
- [ ] If this PR includes major [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note), I have pinged the relevant PM to schedule a changelog post.